pyproject.toml
[project]
name = "boltz-medicine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "biopython>=1.84",
    "boltz>=0.4.1",
    "pyyaml>=6.0.2",
]
